changed Game.WinnerTest() class
Will implement a different winner() function in accordance to the usage
in this test. This change is being made to allow the Game class to
declare multiple winners as the current for loop inside winner() has a
break and only 1 winner is returned (when there may actually be more
than 1 winner). A new loop will be implemented in the main() function to
output the winners